Update keySignatures.md
The insight of viewing the circle of fifths as a line comes from Easley Blackwood, professor emeritus at the University of Chicago.  His insight simplifies understanding key signatures and will prove useful in other chapters on interval identification and key identification in passages that leave the "home" key of a tonal piece.